requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# FastAPI & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Streaming
kafka-python>=2.0.2
redis>=5.0.0
aioredis>=2.0.1

# ML/AI Libraries
ultralytics>=8.0.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Crowd Analysis Specific
scipy>=1.11.0  # For optimization algorithms
filterpy>=1.4.5  # For Kalman filtering
lap>=0.4.0  # For linear assignment problem

# Vector Database
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# LLM & Explainability
openai>=1.3.0
langchain>=0.1.0
anthropic>=0.7.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2
aiofiles>=23.2.1
httpx>=0.25.0

# Monitoring & Logging
prometheus-client>=0.19.0

# Security & Privacy
cryptography>=41.0.0

# System Monitoring
psutil>=5.9.0
# pynvml>=11.5.0  # NVIDIA GPU monitoring (optional)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: GPU acceleration
# tensorrt>=8.6.0  # NVIDIA TensorRT
# onnxruntime-gpu>=1.16.0  # ONNX Runtime GPU
